consolidate defaults in one file
